+++ test/test_runtime_contract.rb
@@ -42,6 +42,7 @@ plasmoid = project_root.join("kde/plasmoids/org.xnix.compatibilitycenter/metadat
   GetPortalAccessPolicy
   GetRuntimeServiceBinding
   GetCompatibilitySettings
+  GetCompatibilitySettingsChangePlan
 ].each do |method|
   assert(contract.include?("name=\"#{method}\""), "D-Bus contract must expose #{method}")
 end